+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 * Python: Added ZRANK command ([#1065](https://github.com/aws/glide-for-redis/pull/1065))
 * Core: Enabled Cluster Mode periodic checks by default ([#1089](https://github.com/aws/glide-for-redis/pull/1089))
 * Node: Added Rename command. ([#1124](https://github.com/aws/glide-for-redis/pull/1124))
+* Python: Added APPEND command ([#1152](https://github.com/aws/glide-for-redis/pull/1152))
 
 #### Features